+++ packages/printer-legacy/src/operation.js
@@ -2,12 +2,15 @@ const {
   graphql: { getTypeName },
 } = require("@graphql-markdown/utils");
 
-const printOperationMetadata = (type) => {
-  let metadata = this.printSection(type.args, "Arguments", {
+const { printSection } = require("./section");
+
+const printOperationMetadata = (type, options) => {
+  let metadata = printSection(type.args, "Arguments", {
     parentType: type.name,
+    parentTypePrefix: options.parentTypePrefix,
   });
   const queryType = getTypeName(type.type).replace(/[![\]]*/g, "");
-  metadata += this.printSection([this.schema.getType(queryType)], "Type");
+  metadata += printSection([options.schema.getType(queryType)], "Type");
 
   return metadata;
 };
